Move map_ids logic into api package
